Crate ntools_utils

Common utility for extended std types

These are left public for convenience.

For example, convenient functions that capitalise a string or extensions to vector functionality are useful everywhere.

Macros

f

Creates a String using interpolation of runtime expressions.

Enums

Error

The error type for ntools_utils

Traits

OptionExt

Extends Option for easy display formatting

SliceExt

Extends functionality for slices of float arrays

SortExt

Extends sorting of vectors and slices by index in-place

StringExt

Extends string types with useful functions

ValueExt

Extends primitives with more specific formatting options